chore: apply review feedback on duplicate code detection
- document npm run duplication in validate and validate:all workflows
- fix jscpd threshold description: upper limit, not a below-current freeze
- rename minToken to minTokens in CODE_QUALITY.md settings table
- extract shared hasExpirationIssue helper in inventory status utils
- move household calculation tests into shared recommendedQuantity tests

Refs: #282
